refactor(ui): extract style converters, formatters, and split components
- Add Impact::label/as_str/FromStr to grey-api
- Introduce ui styles/ module of value->CSS-class converters
- Add formatters/datetime and formatters/incident helpers
- Rename history->probe_history, incidents_timeline->incidents
- Split incident timelines into incident_timeline component
- Extract shared scss base partials (buttons, forms, markdown, page, status)
- Remove dead Status component and example_context_usage
