+++ ui/src/components/_cluster_status.scss
