+++ ui/src/components/_probe_history.scss
