+++ ui/src/components/mod.rs
@@ -1,24 +1,24 @@
 pub mod banner;
 pub mod cluster_status;
 pub mod header;
-pub mod history;
 pub mod icons;
-pub mod incidents_timeline;
+pub mod incident_timeline;
+pub mod incidents;
 pub mod markdown;
 pub mod popover;
 pub mod probe;
+pub mod probe_history;
 pub mod probe_list;
-pub mod status;
 pub mod status_dot;
 pub mod timeline;
 
 pub use banner::{Banner, BannerKind};
 pub use cluster_status::ClusterStatus;
 pub use header::Header;
-pub use history::History;
-pub use incidents_timeline::{IncidentBlock, IncidentsSection};
+pub use incidents::{IncidentBlock, IncidentsSection};
 pub use popover::{Popover, PopoverAlign};
 pub use probe::Probe;
+pub use probe_history::ProbeHistory;
 pub use probe_list::ProbeList;
 pub use status_dot::StatusDot;
 pub use timeline::Timeline;
