+++ ui/src/formatters/mod.rs
@@ -1,7 +1,11 @@
 mod availability;
+mod datetime;
 mod duration;
+mod incident;
 mod si_magnitudes;
 
 pub use availability::*;
+pub use datetime::*;
 pub use duration::*;
+pub use incident::*;
 pub use si_magnitudes::*;
