+++ ui/src/main.rs
@@ -5,6 +5,7 @@ mod components;
 mod contexts;
 pub mod formatters;
 pub mod routes;
+mod styles;
 mod views;
 
 // Export components for both SSR and WASM usage
