+++ ui/src/styles/_forms.scss
@@ -1,0 +1,53 @@
+@use 'variables' as *;
+
+// Create / edit / update forms.
+.incident-form {
+    display: flex;
+    flex-direction: column;
+    gap: 0.75rem;
+    background-color: $background-box;
+    border: 1px solid $color-separator;
+    border-radius: $border-radius-medium;
+    padding: 1.5rem;
+    margin: 0 auto 1.5rem auto;
+
+    h3,
+    h4 {
+        margin: 0;
+        font-weight: 400;
+    }
+
+    label {
+        display: flex;
+        flex-direction: column;
+        gap: 0.25rem;
+        font-size: 0.85rem;
+        color: $color-text-muted;
+    }
+
+    input[type="text"],
+    input[type="datetime-local"],
+    textarea,
+    select {
+        font: inherit;
+        padding: 0.5rem;
+        border: 1px solid $color-separator;
+        border-radius: $border-radius-small;
+        background-color: $background-base;
+        color: $color-text;
+    }
+
+    .incident-form-actions {
+        display: flex;
+        gap: 0.5rem;
+
+        button:not(.primary-button) {
+            cursor: pointer;
+            border: 1px solid $color-separator;
+            background-color: $background-base;
+            color: $color-text;
+            border-radius: $border-radius-small;
+            padding: 0.4rem 0.9rem;
+        }
+    }
+}
