+++ ui/src/styles/_page.scss
@@ -1,0 +1,28 @@
+@use 'variables' as *;
+
+// The full incidents page / per-incident page: lightweight and centred, with no inner card or
+// `.content` chrome of its own.
+.incidents-page {
+    max-width: 980px;
+    width: 100%;
+    margin: 2rem auto;
+    padding: 1.5rem;
+    box-sizing: border-box;
+
+    > h1 {
+        font-weight: 200;
+        margin-top: 0;
+    }
+}
+
+// Muted "nothing to show" text.
+.incidents-empty {
+    color: $color-text-muted;
+    margin: 0;
+}
+
+// Error text shown when a page fails to load.
+.incidents-error {
+    color: $color-status-error;
+    margin: 0;
+}
