+++ ui/src/styles/_status.scss
@@ -1,4 +1,4 @@
-@use '../styles/variables' as *;
+@use 'variables' as *;
 
 // Status indicator and status dot styles
 .status-indicator {
